requirements.txt
ffmpeg-python ~= 0.2.0
json5 ~= 0.9.8
loguru ~= 0.6.0